requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
PyMySQL==1.1.0
